Moving PiFormatter to viz.py
