+++ flamingpy/utils/viz.py
@@ -33,6 +33,7 @@ import numpy as np
 import networkx as nx
 import matplotlib as mpl
 from matplotlib.patches import Patch
+from matplotlib.ticker import Formatter
 import matplotlib.pyplot as plt
 import matplotlib.lines as mlines
 from mpl_toolkits.mplot3d.art3d import Poly3DCollection
@@ -63,7 +64,7 @@ def plot_integer_part(xs, ns, alpha, show=True):
 
     xmin, xmax = alpha * (xs[0] // alpha), alpha * (xs[-1] // alpha) + alpha
     newxticks = np.linspace(xmin, xmax, int((xmax - xmin) // alpha) + 1)
-    ax.xaxis.set_major_formatter(gkp.PiFormatter())
+    ax.xaxis.set_major_formatter(PiFormatter())
 
     plt.plot(xs, ns, ".")
     plt.title("Integer Part")
@@ -86,8 +87,8 @@ def plot_fractional_part(xs, fs, alpha, show=True):
     xmin, xmax = alpha * (xs[0] // alpha), alpha * (xs[-1] // alpha) + alpha
     newxticks = np.linspace(xmin, xmax, int((xmax - xmin) // alpha) + 1)
     newyticks = np.linspace(-alpha / 2, alpha / 2, num=7)
-    ax.xaxis.set_major_formatter(gkp.PiFormatter())
-    ax.yaxis.set_major_formatter(gkp.PiFormatter())
+    ax.xaxis.set_major_formatter(PiFormatter())
+    ax.yaxis.set_major_formatter(PiFormatter())
 
     plt.plot(xs, fs, ".")
     plt.title("Fractional Part")
@@ -110,7 +111,7 @@ def plot_GKP_bins(outcomes, bit_values, alpha, show=True):
 
     xmin, xmax = alpha * (outcomes[0] // alpha), alpha * (outcomes[-1] // alpha) + alpha
     newxticks = np.linspace(xmin, xmax, int((xmax - xmin) // alpha) + 1)
-    ax.xaxis.set_major_formatter(gkp.PiFormatter())
+    ax.xaxis.set_major_formatter(PiFormatter())
 
     plt.plot(outcomes, bit_values, ".")
     plt.title("Binned values")
@@ -142,7 +143,7 @@ def plot_Z_err_cond(hom_val, error, alpha, use_hom_val, show=True):
     print(xmin, xmax, min(val), max(val))
 
     newxticks = np.linspace(xmin, xmax, int((xmax - xmin) // alpha) + 1)
-    ax.xaxis.set_major_formatter(gkp.PiFormatter())
+    ax.xaxis.set_major_formatter(PiFormatter())
 
     plt.plot(val, error, ".")
     plt.xticks(newxticks)
@@ -717,3 +718,43 @@ def draw_decoding(code, ec, dec_objects=None, drawing_opts=None):
     _, ax = syndrome_plot(code, ec, drawing_opts=drawing_opts, index_dict=node_labels)
     if drawing_opts.get("show_recovery"):
         draw_recovery(ax, show_title=drawing_opts.get("show_title"), **dec_objects)
+
+
+def to_pi_string(x, tex: bool = True, d=2):
+    """Convert x, a multiple of sqrt(pi)/2, to a pretty string.
+
+    If x is not a multiple of sqrt(pi)/2, return the unmodified string
+    of x with `d` decimals. If tex is True, add LaTeX $ signs.
+    """
+    remainder = math.remainder(x, np.sqrt(np.pi) / 2)
+    if np.isclose(remainder, 0):
+        integer = round(x / (np.sqrt(np.pi) / 2))
+        pref = int(integer * ((1 - integer % 2) / 2 + integer % 2))
+        x_str = (not bool(round(x))) * "0" + bool(round(x)) * (
+            bool(tex) * "$"
+            + (not bool(1 + pref)) * "-"
+            + bool(1 - abs(pref)) * str(pref)
+            + r"\sqrt{\pi}"
+            + (integer % 2) * "/2"
+            + bool(tex) * "$"
+        )
+        return x_str
+    return "{:.{}f}".format(x, d)
+
+
+class PiFormatter(Formatter):
+    """Formatter for axis-ticks containing multiples of sqrt(pi)/2."""
+
+    def __init__(self, tex: bool = True, d: int = 2):
+        """
+        Initialize the formatter.
+
+        Args:
+            tex: Whether to use LaTeX formatting (i.e. adding $ around the string).
+            d: Number of decimals to use.
+        """
+        self.tex = tex
+        self.d = d
+
+    def __call__(self, x, pos=None):
+        return to_pi_string(x, tex=self.tex, d=self.d)
